requirements.txt
beautifulsoup4>=4.5.3
chardet>=2.3.0
corextopic>=1.0.2
fuzzywuzzy>=0.15.0
hdbscan>=0.8.18
mca>=1.0.3
nilsimsa>=0.3.8
nltk>=3.2.2
numpy>=1.16.0
pandas>=0.19.2
pewtils
polyglot
python-dateutil>=2.6.0
python-Levenshtein
pytz>=2016.10
requests>=2.13.0
scandir>=1.5
scipy>=0.18.1
stopit
tld>=0.7.9
tqdm>=4.11.2
Unidecode>=0.4.20
git+ssh://labsdeploy/pewresearch/pewtils#egg=pewtils
